Add namespace 'wc'
