+++ ruby/oop/wc_command/exec/wc.rb
@@ -1,0 +1,3 @@
+require_relative '../wc'
+
+::Wc::Runner.run
