--- ruby/oop/wc_command/runner.rb
@@ -1,7 +1,0 @@
-class Runner
-  def self.run
-    pp "===Start Wc==="
-
-    Wc.new.run
-  end
-end
