+++ ruby/oop/wc_command/wc/options.rb
@@ -1,12 +1,14 @@
 require 'optparse'
 
-class Wc::Options
-  def self.parse
-    params = {}
-    opt = OptionParser.new
-    opt.on('-l', '--onlylines') { |v| v }
-    opt.parse!(ARGV, into: params)
+module Wc
+  class Options
+    def self.parse
+      params = {}
+      opt = OptionParser.new
+      opt.on('-l', '--onlylines') { |v| v }
+      opt.parse!(ARGV, into: params)
 
-    return params
+      return params
+    end
   end
 end
